Fix warning: implicit conversion changes signedness
visp/modules/core/src/math/matrix/vpMatrix_svd.cpp: In member function ‘void vpMatrix::svdLapack(vpColVector&, vpMatrix&)’:
visp/modules/core/src/math/matrix/vpMatrix_svd.cpp:526:19: warning: conversion to ‘integer’ {aka ‘int’} from ‘unsigned int’ may change the sign of the result [-Wsign-conversion]
  526 |     integer lda = nc;
      |                   ^~
visp/modules/core/src/math/matrix/vpMatrix_svd.cpp:527:19: warning: conversion to ‘integer’ {aka ‘int’} from ‘unsigned int’ may change the sign of the result [-Wsign-conversion]
  527 |     integer ldu = nc;
      |                   ^~
visp/modules/core/src/math/matrix/vpMatrix_svd.cpp:528:38: warning: conversion to ‘int’ from ‘unsigned int’ may change the sign of the result [-Wsign-conversion]
  528 |     integer ldvt = std::min<integer>(nr, nc);
      |                                      ^~
visp/modules/core/src/math/matrix/vpMatrix_svd.cpp:528:42: warning: conversion to ‘int’ from ‘unsigned int’ may change the sign of the result [-Wsign-conversion]
  528 |     integer ldvt = std::min<integer>(nr, nc);
      |                                          ^~
visp/modules/core/src/math/matrix/vpMatrix_svd.cpp:534:77: warning: conversion to ‘int’ from ‘unsigned int’ may change the sign of the result [-Wsign-conversion]
  534 |     integer *iwork = new integer[8 * static_cast<integer>(std::min<integer>(nr, nc))];
      |                                                                             ^~
visp/modules/core/src/math/matrix/vpMatrix_svd.cpp:534:81: warning: conversion to ‘int’ from ‘unsigned int’ may change the sign of the result [-Wsign-conversion]
  534 |     integer *iwork = new integer[8 * static_cast<integer>(std::min<integer>(nr, nc))];
      |                                                                                 ^~

+++ modules/core/src/math/matrix/vpMatrix_svd.cpp
@@ -1,6 +1,6 @@
 /*
  * ViSP, open source Visual Servoing Platform software.
- * Copyright (C) 2005 - 2024 by Inria. All rights reserved.
+ * Copyright (C) 2005 - 2026 by Inria. All rights reserved.
  *
  * This software is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -497,21 +497,21 @@ void vpMatrix::svdLapack(vpColVector &w, vpMatrix &V)
 #else
   {
     vpMatrix U;
-    unsigned int nc = getCols();
-    unsigned int nr = getRows();
+    integer nc = static_cast<integer>(getCols());
+    integer nr = static_cast<integer>(getRows());
 
     if (rowNum < colNum) {
       U = this->transpose();
-      nc = getRows();
-      nr = getCols();
+      nc = static_cast<integer>(getRows());
+      nr = static_cast<integer>(getCols());
     }
     else {
-      nc = getCols();
-      nr = getRows();
+      nc = static_cast<integer>(getCols());
+      nr = static_cast<integer>(getRows());
     }
 
-    w.resize(nc);
-    V.resize(nc, nc);
+    w.resize(static_cast<unsigned int>(nc));
+    V.resize(static_cast<unsigned int>(nc), static_cast<unsigned int>(nc));
 
     double *a = new double[static_cast<unsigned int>(nr * nc)];
     if (rowNum < colNum) {
